remove exectution from runtime_api
