+++ crates/client/rpc/Cargo.toml
@@ -79,6 +79,7 @@ serde = { workspace = true, default-features = true }
 serde_json = { workspace = true, default-features = true }
 serde_with = { workspace = true }
 thiserror = { workspace = true }
+cairo-vm = { workspace = true }
 
 [dev-dependencies]
 rstest = { workspace = true }
