+++ crates/client/rpc/src/utils/mod.rs
@@ -1,4 +1,6 @@
 pub(crate) mod block;
+pub(crate) mod blockifier_state_adapter;
 pub(crate) mod call_info;
+pub(crate) mod execution;
 pub(crate) mod helpers;
 pub(crate) mod transaction;
